Split db pacakge into multiple files
